[POM-80] Refactor Attributes
